refactor(views): fix profile links rendering and complete meme view structure
- Pass targetEl to renderLinks in profile view
- Hide socials container if no links present
- Remove stale ui/ imports, mount chat from src/chat.js
- Add meme view with shell, grid, card, controls
