+++ src/ads/load.js
@@ -1,7 +1,7 @@
 import { ADS_CACHE_KEY, ADS_CACHE_MS, JUP_SWAP, EXPLORER, FALLBACK_LOGO, shortAddr } from "../config/env.js";
 import { getJSON } from "../utils/tools.js";
 import { normalizeSocial, iconFor } from "../data/socials.js";
-import { normalizeWebsite } from "../utils/normalize.js";
+import { normalizeWebsite } from "../data/normalize.js";
 
 function readInlineAds(){
   try{
