--- src/ui/page.js
@@ -1,1 +1,0 @@
-
