+++ src/views/meme/legal.js
